+++ core/interfaces.go
@@ -35,6 +35,7 @@ type Administration interface {
 	AdmGetTestModel() string
 	AdmCreateGlobalConfig(setting string) (*model.GlobalConfig, error)
 	AdmGetGlobalConfig() (*model.GlobalConfig, error)
+	AdmUpdateGlobalConfig(config *model.GlobalConfig) error
 }
 
 type administrationImpl struct {
@@ -54,6 +55,9 @@ func (s *administrationImpl) AdmCreateGlobalConfig(setting string) (*model.Globa
 }
 func (s *administrationImpl) AdmGetGlobalConfig() (*model.GlobalConfig, error) {
 	return s.app.admGetGlobalConfig()
+}
+func (s *administrationImpl) AdmUpdateGlobalConfig(config *model.GlobalConfig) error {
+	return s.app.admUpdateGlobalConfig(config)
 }
 
 //Encryption exposes APIs for the Encryption building block
@@ -87,6 +91,7 @@ type Storage interface {
 	SetStorageListener(storageListener StorageListener)
 	CreateGlobalConfig(setting string) (*model.GlobalConfig, error)
 	GetGlobalConfig() (*model.GlobalConfig, error)
+	SaveGlobalConfig(config *model.GlobalConfig) error
 }
 
 //StorageListener listenes for change data storage events