+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+
 ### Added
+- Expose get global config admin API [#36] 
 - Expose get global config admin API [#35] (https://github.com/rokwire/core-building-block/pull/43)
 - Set up auth framework. [#16] (https://github.com/rokwire/core-building-block/issues/16)
 - Expose version API. [#13] (https://github.com/rokwire/core-building-block/issues/13)